Update utility composite node, update order preview actor, clean up functionality of gameplay ability. Switch to softobjectpointer where needed

+++ Source/ART/Public/ARTCharacter/ARTCharacterBase.h
@@ -59,7 +59,7 @@ public:
 
 	//Team ID, handled by Interface, different ID are hostile except +50 ID as Neutral
 	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "ART|Character")
-	int TeamNumber;
+	uint8 TeamNumber;
 
 	FGenericTeamId GetGenericTeamId() const override;
